add function and fix
