+++ package.json
@@ -24,6 +24,7 @@
     "lodash.throttle": "^4.1.1",
     "modern-normalize": "^1.1.0",
     "notiflix": "^3.2.6",
+    "overlayscrollbars": "^2.2.0",
     "simplelightbox": "^2.14.1",
     "swiper": "^9.4.1",
     "tui-pagination": "^3.4.1"
