+++ src/js/categories.js
@@ -1,25 +1,51 @@
-// import { getTopBooksByCategories } from "./top-books";
-// import { useBooksApi } from '../services/booksApi';
+import { getTopBooksByCategories } from "./top-books";
+import { renderBooks } from "./renderBooksByCategories";
 
-// const categoryAll = document.querySelector('.category-link');
-// console.log(categoryAll);
+const categoryAll = document.querySelector('.category-link');
 
-// const booksApi = useBooksApi()
-// categoryAll.addEventListener('click', () => {
-//     booksApi
-//         .getTopBooks()
-//         .then((res) => {
-//             const topBooksByCategories = res.reduce((acc, categoryBooks) => {
-//                 const category = categoryBooks.category;
-//                 const topBooks = categoryBooks.books.slice(0, 5);
-//                 acc[{ category }] = topBooks;
-//                 return acc;
-//             }, {});
-//             getTopBooksByCategories(topBooksByCategories);
-//         }).catch((err) => {
-//             console.log(err);
-//         })
-//         .finally(() => {
-//         });
+
+const categoryLinksElements = document.querySelectorAll('.category-link');
+
+export function getCategoryElements() {
+
+    categoryLinksElements.forEach((link) => {
+        link.addEventListener('click', (evt) => {
+            evt.preventDefault();
+            categoryLinksElements.forEach((el) => el.classList.remove('active'));
+            link.classList.add('active');
+        })
+    });
+}
+categoryAll.addEventListener('click', () => {
+    getTopBooksByCategories();
+
+})
+// categoryLinksElements.addEventListener('click', () => {
+//     renderBooks(category);
+
 // })
+// для отримання списку книг по одній категорії
 
+// function showBooks(category) {
+//     const bookList = document.getElementById('bookList');
+
+//     // Очищаємо список книг
+//     bookList.innerHTML = '';
+
+//     // Фільтруємо книги за вибраною категорією
+//     const filteredBooks = books.filter(book => book.category === category);
+
+//     if (filteredBooks.length === 0) {
+//         // Якщо книг у вибраній категорії немає, показуємо повідомлення
+//         const message = document.createElement('li');
+//         message.textContent = 'Книг у цій категорії не знайдено.';
+//         bookList.appendChild(message);
+//     } else {
+//         // Якщо є книги у вибраній категорії, відмальовуємо їх
+//         filteredBooks.forEach(book => {
+//             const listItem = document.createElement('li');
+//             listItem.textContent = book.title;
+//             bookList.appendChild(listItem);
+//         });
+//     }
+// }
